+++ src/test/java/com/studyolle/WithAccountSecurityContextFactory.java
@@ -1,7 +1,7 @@
 package com.studyolle;
 
 import com.studyolle.account.AccountService;
-import com.studyolle.account.SignUpForm;
+import com.studyolle.account.form.SignUpForm;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
